% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leninru.R
\docType{data}
\name{leninru_volumes}
\alias{leninru_volumes}
\title{Titles & descriptions of Lenin's Collected Works}
\format{
A data frame, containing three columns:
\describe{
\item{vol}{
Volume number of complete works. (Number between 1 and 55)
}
\item{title}{Title of volume}
\item{description}{
Description of works within the volume and time periods covered.
}
}
}
\source{
\url{https://leninism.su/works}
}
\usage{
leninru_volumes
}
\description{
All works written by Lenin available on leninism.su
}
\details{
The text is stored as a data frame with two columns. Each row corresponds to
one of 55 volumes, and provides the title of the volume and a description of
the volume written by the publisher.
}
\keyword{datasets}
